requirements.txt
django>=4.2
djangorestframework
django-filter
drf-spectacular
psycopg2-binary
dj-database-url
